chore(cardio): extract assembleCardioData to a shared, client-free module
Splits the read-path assembler out of `lib/data/cardio.ts` (which
imports the 'use client' browser Supabase client) into a new
`lib/data/cardio-shared.ts` that imports neither client. Now the
server reader (`cardio-server.ts`) doesn't transitively pull the
browser client into the server bundle, and the browser entry doesn't
pull the server client. Both files become thin wrappers that hand
their respective Supabase clients to the shared assembler.

Also documents the loose `SupabaseClient` parameter type — narrowing
to `SupabaseClient<Database>` is gated on adopting
`supabase gen types typescript`, which isn't in scope here.
